// src/app/shared/models/tag.interface.ts
export interface TagInterface {
  created: string;
  name: string;
  company: string;
  advertiser: string;
  publisher: string;
  browserStatus: string;
  browser: [];
  deviceTypeStatus: string;
  deviceType: [];
  versionStatus: string;
  version: [];
  countryStatus: string;
  country: [];
  adServerUrl: string;
  statType: string;
  // subid: string;
  // limit: string;
  subids: [];
  rotationType: string;
  url: string;
  params: [];
  _id: string;
  _key: string;
  _rev: string;
}